bodhibot-backend/chat/views.py
from django.shortcuts import render
from rest_framework.permissions import IsAuthenticated
from rest_framework.views import APIView
from rest_framework.response import Response
from .models import Chat, Message
from .serializers import ChatSerializer, MessageSerializer

class UserChatView(APIView):
    permission_classes = [IsAuthenticated]

    def get(self, request):
        chats = Chat.objects.filter(user= request.user).order_by('-last_message')
        serializer = ChatSerializer(chats, many= True)
        return Response(serializer.data)
    
    