bodhibot-backend/requirements.txt
accelerate==1.8.1
amqp==5.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1749054771823/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
celery==5.5.3
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
channels==4.2.2
channels_redis==4.2.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
conda @ file:///croot/conda_1749171252582/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1744662537425/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1748537199219/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
constantly==23.10.4
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1748977429425/work
cycler==0.12.1
daphne==4.2.1
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
fonttools==4.58.5
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
hyperlink==21.0.0
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
incremental==24.7.2
Jinja2==3.1.6
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
kiwisolver==1.4.8
kombu==5.5.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1738943416351/work
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging @ file:///croot/packaging_1734472117206/work
peft==0.16.0
pillow==11.3.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///croot/pygments_1744664109463/work
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rich @ file:///croot/rich_1732638981168/work
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
safetensors==0.5.3
scipy==1.16.0
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.53.2
triton==3.3.1
truststore @ file:///croot/truststore_1736550121485/work
Twisted==25.5.0
txaio==25.6.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
zope.interface==7.2
zstandard @ file:///croot/zstandard_1731356346222/work
